logs ok

+++ app/contracts/app.sol
@@ -6,7 +6,7 @@ contract Votrice {
 // voters structure
 struct Voter {
     bool voted; // has voted ?
-    uint vote; // index of the choice
+    uint8 vote; // index of the choice
 }
 
 // choices stucture
@@ -22,13 +22,13 @@ mapping(address => Voter) public voters;
 Choice[] public choices;
 
 // constructor
-function Votrice(uint _numChoices) public {
+function Votrice(uint8 _numChoices) public {
     chairperson = msg.sender;
     choices.length = _numChoices;
 }
 
 // function to vote for someone
-function vote(uint myChoice) public {
+function vote(uint8 myChoice) public {
     Voter storage sender = voters[msg.sender];
     require(!sender.voted && myChoice <= choices.length);
     sender.voted = true;
@@ -37,9 +37,9 @@ function vote(uint myChoice) public {
 }
 
 // function to get the most voted choice
-function getWinningChoice() public constant returns (uint winningChoice) {
-    uint winningCount = 0;
-    for (uint j = 0; j < choices.length; j++) {
+function getWinningChoice() public constant returns (uint8 winningChoice) {
+    uint256 winningCount = 0;
+    for (uint8 j = 0; j < choices.length; j++) {
         if (choices[j].count > winningCount) {
             winningCount = choices[j].count;
             winningChoice = j;
